feat: wrap all elements in a functional vue component
BREAKING CHANGE: All elements are now Vue components, in some cases where a ref is used to get the
nativeView will have to be updated to use `this.$refs.element.$el.nativeView` instead of
`this.$refs.element.nativeView`. There may be other breaking changes this introduces that are not
known at this point.

re #266 re #241