fix: make hooks compatible with CLI's 6.0.0 changes
In NativeScript CLI 6.0.0 release the bundle workflow will be the only one available. Currently, the nativescript-vue has some hooks that verify bundle is passed. There's no need of them for 6.0.0 release.
Implement a check based on the CLI's version and skip the validation in case CLI is 6.x.x or newer.